Add build script and DIB element for Fedora image
This enables us to build an image with the Heat agents, which are
needed to use the OS::Heat::SoftwareDeployment Heat resource.
Magnum's certificate revocation requires the SoftwareDeployment
resource to update the certificate without triggering an
OS::Nova::Server rebuild.

This will require some driver changes.  I will submit a driver that
works with this image in the next patchset.

Co-Authored-By: Randall Burt <randall.burt@rackspace.com>
blueprint revoke-cluster-cert
Change-Id: Ifcf6196ee851c289e1fd9ed2ecb908ed90deb3f2